+++ src/client/App.svelte
@@ -17,7 +17,8 @@
 	import SourceMetaBuildTreeExplorer from './SourceMetaBuildTreeExplorer.svelte';
 	import SourceMetaTreeExplorer from './SourceMetaTreeExplorer.svelte';
 	import SourceMetaTreeExplorers from './SourceMetaTreeExplorers.svelte';
-	import {createSourceTree, SourceTree} from './sourceTree.js';
+	import {createSourceTree} from './sourceTree.js';
+	import type {SourceTree} from './sourceTree.js';
 	import type {ProjectState} from '../server/projectState.js';
 	import type {View} from './view.js';
 	import {provideProjectState} from './projectState.js';