+++ src/fs/importTs.ts
@@ -2,7 +2,7 @@ import lexer from 'es-module-lexer';
 
 import {paths, TS_EXTENSION} from '../paths.js';
 import {randomInt} from '../utils/random.js';
-import {createSwcBuilder} from '../build/swcBuilder.js';
+import {createEsbuildBuilder} from '../build/esbuildBuilder.js';
 import {BuildConfig} from '../config/buildConfig.js';
 import type {BuildContext, TextBuildSource} from '../build/builder.js';
 import {DEFAULT_ECMA_SCRIPT_TARGET} from '../build/tsBuildHelpers.js';
@@ -52,7 +52,7 @@ export const importTs = async (
 		dev: true,
 		sourceMap: false,
 		target: DEFAULT_ECMA_SCRIPT_TARGET,
-		// TODO these last two aren't needed, maybe the swc compiler's type should explicitly choose which options it uses?
+		// TODO these last two aren't needed, maybe the builder's type should explicitly choose which options it uses?
 		servedDirs: [],
 		externalsAliases: {},
 		state: {},
@@ -80,7 +80,7 @@ const compileFileAndImports = async (
 		dirBasePath: stripStart(dir, paths.source),
 		extension: TS_EXTENSION,
 	};
-	const builder = createSwcBuilder();
+	const builder = createEsbuildBuilder();
 	const {
 		builds: [build],
 	} = await builder.build(source, buildConfig, ctx);