+++ src/index.ts
@@ -2,6 +2,7 @@ import './globalTypes.js'; // makes global types available to user code
 
 export {test, t} from './oki/oki.js';
 
-export {Task, TaskError} from './task/task.js';
+export type {Task} from './task/task.js';
+export {TaskError} from './task/task.js';
 
-export {Gen} from './gen/gen.js';
+export type {Gen} from './gen/gen.js';